feat(bid-document): proper error handling
